separate file for results classes Feature(s)
