+++ src/popsift/sift_pyramid.h
@@ -10,10 +10,11 @@
 #include <iostream>
 #include <vector>
 
-#include "s_image.h"
 #include "sift_conf.h"
-#include "sift_extremum.h"
 #include "sift_constants.h"
+#include "features.h"
+
+#include "s_image.h"
 #include "sift_octave.h"
 
 namespace popsift {
